drop retry.WithWaitLabel/WaitLabelFromContext instead of deprecating
The wait-label context helpers only shipped on main a week ago in the
sync-stats work (#1004) and have no external adopters; rather than carry
deprecated re-export shims in pkg/retry, remove them and point all
callers at the canonical pkg/ratelimit versions.
